fix(core): read a sealed secret's key_id at the width its column declares (ADR-018 Inc.1)
Found while verifying ADR-097 Inc.4 on the lab core: creating one notification
channel made core log this every 30 seconds, forever.

  WARN failed to load notification channels
  error=... decoding column "key_id": Rust type `i64` (as SQL type `INT8`)
  is not compatible with SQL type `INT4`

sqlx checks a column's type when the row is decoded, so reading an INTEGER
column into i64 fails the whole row -- and load_routing degrades a failed read
to an empty snapshot. A deployment with a notification channel therefore loaded
zero channels and delivered no notifications at all, with a WARN line as the
only sign.

Three of the nine sealed-secret readers were wrong the same way, and each one
silently disables a whole feature the moment a row exists:

  notifications.rs   notification_channels (0009, INTEGER)  -> all delivery
  oidc.rs            oidc_providers        (0044, INTEGER)  -> SSO sign-in
  forward_store.rs   forward_destinations  (0050, INTEGER)  -> tee credentials

The six correct ones stay as they are: credentials is BIGINT (so i64 is right
there), and rca / ldap_config / web_tls / bus_tls / bus_callout are already i32.

None of this was reachable by a unit test -- they are live PostgreSQL reads --
and the lab held zero rows in all three tables, which is exactly what ADR-097's
own remnant said about notification channels. It stood for eight releases.

The guard derives the rule instead of restating it: it counts the key_id widths
declared in migrations/, pins that exactly one is BIGINT, then reads the whole
crate through srcread and refuses an i64 read anywhere but that one file. Two
floors, because the healthy answer here is "found nothing" and that has to be
distinguishable from "looked at nothing". Broken in both directions: putting the
i64 back fails the test, naming the file and the line.

Workspace 2,790 -> 2,791.

+++ crates/yagra-core/src/forward_store.rs
@@ -270,7 +270,7 @@ impl ForwardStore {
     }
 
     fn row_to_open(&self, row: sqlx::postgres::PgRow) -> anyhow::Result<OpenDestination> {
-        let key_id: Option<i64> = row.try_get("key_id")?;
+        let key_id: Option<i32> = row.try_get("key_id")?;
         let secret = match key_id {
             None => None,
             Some(key_id) => {
